Expose the generic tuple store and improve parsing of .nt files
Before, nomunofu only accepted 3-tuple items. Now, it configurable on
the commandline.  Mind the fact, that there is no check that are done,
a priori, so you have to call the correct command line option whether
it is serve or index depending on the database and the .nt input file.

If you pass the wrong N number, it will fail, in non-obvious way.

The make targets are updated to reflect that the test file contains
3-tuple items.

Also:

- Parse espaced double quotes in strings;

- Decimal and doubles are parsed and stored as doubles (see the commit
  before the previous commit 27a5cf26ad022f9e6b61d36cd50d788d21d13c0e);

- opengis points are stored as strings;

- mathml is stored as a string;

- "foobar"@LANG is stored as list ("fooobar" LANG) instead of the
  nomunofu/lang/LANG indirection.  That is faster to import, it will
  be up to the client to translate the sparql query into the correct
  object filter;

The code is cleaner.